Update MVP superclasses to TypeScript
