--- src/client/superclasses/Model.js
@@ -1,7 +1,0 @@
-/** Model superclass */
-export default class Model {
-  /** @param {EventEmitter} emitter */
-  constructor(emitter) {
-    this.emitter = emitter;
-  }
-}
