+++ src/client/superclasses/Model.ts
@@ -1,0 +1,9 @@
+import * as EventEmitter from 'eventemitter3';
+
+export default class Model {
+  protected readonly emitter: EventEmitter;
+
+  public constructor(emitter: EventEmitter) {
+    this.emitter = emitter;
+  }
+}
